+++ dwds/debug_extension_mv3/web/chrome_api.dart
@@ -129,6 +129,7 @@ class Target {
 }
 
 /// chrome.tabs APIs
+/// chrome.storage APIs
 /// https://developer.chrome.com/docs/extensions/reference/storage
 
 @JS()