+++ dwds/CHANGELOG.md
@@ -4,6 +4,10 @@
   Dart Debug Extension notifying that the Dart app is ready.
 - Include an optional param to `Dwds.start` to indicate whether it is running
   internally or externally.
+- Fix null cast error on expression evaluations after dwds fails to find class
+  metadata.
+- Include the entire exception description up to the stacktrace in
+  `mapExceptionStackTrace`.
 
 ## 16.0.1